+++ src/pub/app.js
@@ -11,6 +11,7 @@ import Overview from './overview.js';
 import Config from './config.js';
 import Settings from './settings.js';
 import Tracer from './tracer.js';
+import Whitelist_ips from './whitelist_ips.js';
 import {Logs, Dock_logs} from './logs.js';
 import {Enable_ssl_modal} from './common.js';
 import React from 'react';
@@ -29,6 +30,39 @@ const App = withRouter(class App extends Pure_component {
     componentDidMount(){
         const _this = this;
         this.etask(function*(){
+            const version = yield ajax.json({url: '/api/version'});
+            setdb.set('head.version', version.version);
+        });
+        this.etask(function*(){
+            this.on('uncaught', e=>console.log(e));
+            const mode = yield window.fetch('/api/mode');
+            let block_ip;
+            if (block_ip = mode.headers.get('x-lpm-block-ip'))
+            {
+                setdb.set('head.blocked_ip', block_ip);
+                return _this.props.history.replace('/whitelist_ips');
+            }
+            _this.load_data();
+            const data = yield mode.json();
+            if (data.logged_in)
+            {
+                if (!['/login', '/whitelist_ips'].includes(
+                    _this.props.location.pathname))
+                {
+                    return;
+                }
+                return _this.props.history.replace('/overview');
+            }
+            if (_this.props.location.pathname=='/login')
+                return;
+            return _this.props.history.replace({
+                pathname: '/login',
+                search: _this.props.location.search,
+            });
+        });
+    }
+    load_data = ()=>{
+        this.etask(function*(){
             const locations = yield ajax.json({url: '/api/all_locations'});
             locations.countries_by_code = locations.countries
             .reduce((acc, e)=>({...acc, [e.country_id]: e.country_name}), {});
@@ -38,10 +72,6 @@ const App = withRouter(class App extends Pure_component {
             const settings = yield ajax.json({url: '/api/settings'});
             setdb.set('head.settings', settings);
             window.ga('set', 'dimension1', settings.customer);
-        });
-        this.etask(function*(){
-            const version = yield ajax.json({url: '/api/version'});
-            setdb.set('head.version', version.version);
         });
         this.etask(function*(){
             const version = yield ajax.json({url: '/api/last_version'});
@@ -74,28 +104,13 @@ const App = withRouter(class App extends Pure_component {
             const warnings = yield ajax.json({url: '/api/warnings'});
             setdb.set('head.warnings', warnings.warnings);
         });
-        this.etask(function*(){
-            this.on('uncaught', e=>console.log(e));
-            const data = yield ajax.json({url: '/api/mode'});
-            if (data.logged_in)
-            {
-                if (_this.props.location.pathname!='/login')
-                    return;
-                return _this.props.history.replace('/overview');
-            }
-            if (_this.props.location.pathname=='/login')
-                return;
-            return _this.props.history.replace({
-                pathname: '/login',
-                search: _this.props.location.search,
-            });
-        });
-    }
+    };
     render(){
         return <div className="page_wrapper">
               <Enable_ssl_modal/>
               <Switch>
                 <Route path="/login" exact component={Login}/>
+                <Route path="/whitelist_ips" exact component={Whitelist_ips}/>
                 <Route path="/dock_logs" exact component={Dock_logs}/>
                 <Route path="/" component={Page}/>
               </Switch>